Small refactors in generated content
