+++ src/setup/steps/add-content/add-content.ts
@@ -2,11 +2,11 @@ import { promises as fs } from "fs"
 import { exitWithError } from "../../../helpers/exit-with-error"
 import { commandInstance } from "../../../instance"
 import { Step } from "../../step"
+import { generatePage } from "./components/generate-page"
+import { generateWithDefaultGlobalStyles } from "./components/generate-with-default-global-styles"
 import { generateApp } from "./generate-app"
 import { generateIndex } from "./generate-index/generate-index"
-import { indexCSSModule } from "./generate-index/with-css-modules/generate-index-module"
-import { generatePage } from "./generate-page"
-import { generateWithDefaultGlobalStyles } from "./generate-with-default-global-styles"
+import { indexCSSModule } from "./generate-index/with-css-modules/index-css-module"
 import { globalStyles } from "./global-styles"
 
 export const addContentStep: Step = {
